requirements.txt
litellm
kornia==0.7.2
diffusers==0.25.1
transformers==4.37.2
matplotlib==3.8.3
numpy==1.24.4
torchmetrics==1.3.1
tqdm==4.65.0
cprint==1.2.2
huggingface_hub==0.20.3
accelerate==0.27.2
tyro==0.7.3
ninja==1.11.1.1
einops==0.7.0
vispy==0.14.1
moviepy==1.0.3
trimesh[easy]==4.1.4
jaxtyping==0.2.25
omegaconf==2.3.0
viser==0.1.21
sentence_transformers==2.3.1
iopath==0.1.9
plyfile
torchtyping
cython
lpips
gdown
yacs
